pyproject.toml
[tool.poetry]
name = "playwhat"
version = "0.1.0"
description = ""
authors = ["0xChloe <chloe@catastrophesoftware.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
flask = "^3.0.2"
bs4 = "^0.0.2"
python-dotenv = "^1.0.1"
requests = "^2.31.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
